src/requirements.txt
Flask==0.10.1
requests==2.4.3
docker-py==1.8.1
PyMySQL==0.6.4
psycopg2==2.6.1
gunicorn==19.3.0
python-decouple==3.0
